refactor: change directory of auth validations
